pyproject.toml
[tool.poetry]
name = "personal portfolio"
version = "0.2.0"
description = "Fully responsive personal portfolio single-page website, responsive for all devices, built using HTML, CSS, JavaScript and Python."
authors = ["ivansaul <ihuamanis@uni.pe>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
jinja2 = "^3.1.4"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.9"
python-semantic-release = "^9.10.1"
pytest = "^8.3.3"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
preview = true
select = ['E', 'F', 'W', 'I']
ignore = ["E501"]
